+++ hc256.cpp
@@ -107,7 +107,7 @@ void HC256Policy::OperateKeystream(KeystreamOperation operation, byte *output, c
 		//  writing the result to the output buffer. In all cases the
 		//  keystream is written to the output buffer. The optional part is
 		//  adding the input buffer and keystream.
-		if ((operation & INPUT_NULL) != INPUT_NULL)
+		if ((operation & static_cast<int>(INPUT_NULL)) != static_cast<int>(INPUT_NULL))
 		{
 			xorbuf(output, input, BYTES_PER_ITERATION);
 			input += BYTES_PER_ITERATION;
